Add a no-op OPENSSL_INIT_NO_ATEXIT
Upstream OpenSSL will, by default, register an atexit callback to free
globals, even though the process is going to be destroyed anyway.

Not only is this pointless, but it introduces UAFs on race condition if
some thread outlives the main thread. Some projects have started to
notice this, and use OPENSSL_INIT_NO_ATEXIT from 1.1.1b. Add the
constant for compatibility.

We already do not call atexit, so the option is a no-op.

Change-Id: Ia17c529c27646507100ebb69c884e2db9cb70431