Finish documenting x509.h
It took 3.5 years, but this header is now DONE! I opted to add a section
for each extension just because there were so many functions. It's a
little weird because, for example, we don't have a section for key usage
because it's just BIT STRING. But I think this is better than having a
great big "types for built-in extensions" section.

Fixed: 426
Change-Id: Ifc7684cc6ff6a211ea1f5065eff67663adf004b3