Add ERR_lib_symbol_name and ERR_reason_symbol_name
CPython needs this operation. See
https://github.com/openssl/openssl/issues/19848 and
https://discuss.python.org/t/error-tables-in-the-ssl-module/25431 for
details.

In principle, our functions already return the symbol names. The
differences are:

- Our library strings say "common libcrypto routines" instead of
  "CRYPTO".
- The global reason codes say "internal error" instead of
  "INTERNAL_ERROR". (We should consider changing this.)
- The library forwarding reason codes (ERR_R_BN_LIB) say the library
  string instead of "BN_LIB". (We should consider changing this.)
- errnos report strerror
- Unknown errors return "unknown error" because we've found that
  projects tend to crash when these APIs return NULL.

The new APIs consistently return the symbol name, when available. If
unavailable (ERR_LIB_SYS's errno reasons), it returns NULL because I
assume callers would rather be able to handle that case themselves.
Hopefully this will not be as common so callers can take on this one.

Change-Id: Idd9e4b1cb5a4f64513310d8066d6bf3970722c23