Tidy up Rust HPKE binding.
The doc test didn't compile and needs support for generating keys in
order to be a good example. Also bind all the AEADs that we support, and
have the encapsulated key be a second return value when creating a
sender, rather than carrying it in memory for the whole lifetime.

Change-Id: I1533560a925d4e239eedbfa0f4213d9e79085b77