Embed resources directly into runner
This removes the need to ship the PEM files with the built runner.
Instead we can use go:embed to pick up the key files. We do, annoyingly,
need to write the Channel ID file to a temporary, but it's not a huge
deal. When/if we rework all this to JSON, we can avoid this.

Change-Id: Ie0d187a5396546dc157906430639c26b3cc59ca2