runner: Remove the ability to configure multiple certificates
While we do want to add this to the shim and real TLS stack, we have no
need to configure this on the runner. This is a remnant of runner being
a production TLS stack. We only ever configure one certificate because
we know what we expect the peer to send.

Change-Id: Idaafc66a2a7f83a929a6a61c19b97cdef158c534