Mark spx exports from experimental as deprecated.
Since we are saying this will die when standardized, let us
mark this as deprecated.

We then selectively allow it in the speed tool and in our tests.

Users picking up this function "just because" will then recieve
the (accurate) deprecation warning about it's use. They can then
suppress the warning if they are fine with using an experimental
api, and if they do we are well within reasonable grounds to say
"you were using an experimental API, you fix your own code please"
when this standardizes and changes.

If we like this approach, I'll go back and apply it to kyber
(which will have some other fallout consequences to fix) but this
one should be painless right now.

This can also be applied to Dilithium when it comes back.

Future experimental's could be added in this manner.

Change-Id: Ie3b41cf16278868562ef1c8b28f2caed5e0e2dd1