Flatten crypto/fipsmodule/CMakeLists.txt up a layer
This is one step towards flattening the libcrypto build into the
top-level CMakeLists.txt file. (This is to align the CMake build with
our other builds, and also make it easier to consume a CRYPTO_SOURCES
variable without having to prepend "../" in front of each value.)

This also avoids a rather annoying corner of CMake: files generated in a
different directory don't work well, which is why we had all this mess
with EXTERNAL_OBJECT, GENERATED, and bcm_o_target. Globbing it into one
file is a bit unwieldy, but avoids this. (CMake is incredibly bad at
custom rules.)

Bug: 542
Change-Id: Ia5038511af339a0eae2af56875a42581eb1ed15b