Use sources.json to find pki_test data in run_android_tests.go
We now have (some of) our test sources in an easily parseable form.
run_android_tests.go no longer needs to crawl the source tree.

Note this required fixing the .gitignore rules. If a .gitignore line
doesn't have a slash at the start or middle, it applies to
subdirectories as well. This is confusing, so I just stuck a leading
slash in front of all of them.

Bug: 681
Change-Id: I389c2a0560594fbd23c60b5b614b0ccfedf28926