Prioritize spec compliance over a 10 year old f5 bug.
This mirrors a change being done in LibreSSL.

This hack was for an f5 bug that was fixed in 2014. Libre kept
the same fix in 2017 because as of that time there were some f5
boxes that had not been updated in three years.

As of today the support article for F5 relating to this bug is
archived on their web site:

https://support.f5.com/csp/article/K37345003

This seems top indicate tht f5 no longer supports caring about this

So let's conform to the spec instead of worrying about load balancers
that have not had software updates applied in 10 years.

Change-Id: Ia5cdf8006307dc56e23c1dc72bcc901e0ff64a53