Reset DTLS1_BITMAP without resorting to memset
After f94f3ed3965ea033001fb9ae006084eee408b861, there's a std::bitset in
it, so memset is not guaranteed to do what we want.

There are still other problems here, and we may decide not to keep the
reset at all, but fix the memset for now. See these discussions for
more details:
https://boringssl-review.googlesource.com/c/boringssl/+/68027/comment/7ff3d697_a6808cee/
https://boringssl-review.googlesource.com/c/boringssl/+/68027/comment/74cabada_eadf5394/

Change-Id: I79c62d71d74a58be4ac822e0cc5a41586472a26a