Initial commit of libcrux ml-kem
This change adds the verified libcrux ml-kem code into the third_party
directory. There are two versions, portable and avx2.
An optimized, but unverified, impelmentation of SHA-3 is provided in
there as well. ML-KEM performance heavily depends on SHA-3 and a 4x API
is used in some cases, which is not generally useful.

This is NOT updated to the final FIPS 203 spec yet.

A local build script is provided for convenience.

Change-Id: I4fbcb199ea145e34bd3687e7e93e870a5f074ad6