Initial commit of libcrux ml-kem
This change adds the libcrux ml-kem 768 code into the third_party
directory and integrates it into mlkem/mlkem.c.

A portable and an avx2 variant are included.

This code includes a few manual changes that are automated
over the next couple of days.
Tests and speed tests have been added as well.

Change-Id: I4fbcb199ea145e34bd3687e7e93e870a5f074ad6