Make googletest a full dependency, not a dev_dependency
In principle, googletest is only needed for tests and could be a
dev_dependency. But it would then be impossible for downstream modules
to run BoringSSL's tests. The downstream module provides the toolchain,
so we may need to request they run tests when triaging issues. If
https://github.com/bazelbuild/bazel/issues/22187 is ever fixed, we can
change this.

With this change, the following works:

  cd util/bazel-example
  bazelisk test @boringssl//:crypto_test

Change-Id: Ied2276047de134883d6b61b0789f3c7bfcaad669