Move urandom and OS crypto support out of BCM
BCM uses only passive entropy, in that from BCM's
point of view, entropy is requested and obtained
from an external source, and fed to BCM's CTR_DRBG
as required.

We introduce an interface boundary defined in
bcm_interface.h between BCM and the rest of libcrypto.
the actual gathering of entropy in an OS specific
manner is handled outside of BCM by libcrypto, while
the CTR_DRBG remains in BCM. This includes the
introduction of a bcm_status type to be used as the
return value for functions inside BCM called from the rest
of libcrypto.

This changes BCM to use only BCM_RAND_bytes instead of
the public RAND_bytes, so that BCM no longer calls
any public RAND functions directly, and moves the
public RAND functions not used in BCM outside to
rand_extra.

Bug: 723

Change-Id: I6b618dfe4df257f67971e88cbd79126c837e21d6