Add tests for some odd escaping behavior in the CONF parser
Honestly, these are probably bugs, but add tests for them so, if we
change the behavior, we do so intentionally. The escape processing logic
has checks for hitting the end of the string early.

At first, I had a hard time reaching this case because this is normally
processed as a line continuation. However, the line continuation logic
is not escape-aware. It just assumes if your line ends "...\\\\", that
the last backslash is not a continuation. I.e. it doesn't correctly
count escapes.

This is almost certainly a bug, but means the escape + EOF behavior is
reachable. Even more interesting is that it seems to intentionally
terminate quote handling. Add tests for these cases.

Change-Id: I9e058ec2b1ce3e20d87eab28a64c79880e3e5cae