Namespace crypto/test/file_util.h
We've historically not wrapped the test bits in namespaces, since it's
not general library code or anything. (Possibly we should?)

Unfortunately Android's libbase ended up leaving a TemporaryFile class
in the root namespace. That, which actually is general library code,
really should have been namespaced, but moving it will take a while, so
namespace ours.

Change-Id: I600e2dcf347e4cc6b4b81be5a26659698fe80064